Moving the AccountsApiClient and associated config into the shared project. Fixing references in existing places.
