+++ src/SFA.DAS.EmployerIncentives.Api/Startup.cs
@@ -17,6 +17,7 @@ using SFA.DAS.EmployerIncentives.Application.Queries.EligibleApprenticeshipsSear
 using SFA.DAS.EmployerIncentives.Configuration;
 using SFA.DAS.EmployerIncentives.Infrastructure;
 using SFA.DAS.SharedOuterApi.AppStart;
+using SFA.DAS.SharedOuterApi.Configuration;
 using System.Collections.Generic;
 
 namespace SFA.DAS.EmployerIncentives.Api
