+++ src/SFA.DAS.EmployerIncentives/Application/Services/AccountsService.cs
@@ -1,9 +1,10 @@
 ﻿using System;
 using System.Threading.Tasks;
-using SFA.DAS.EmployerIncentives.Configuration;
 using SFA.DAS.EmployerIncentives.InnerApi.Requests.Accounts;
 using SFA.DAS.EmployerIncentives.InnerApi.Responses.Accounts;
 using SFA.DAS.EmployerIncentives.Interfaces;
+using SFA.DAS.SharedOuterApi.Configuration;
+using SFA.DAS.SharedOuterApi.Interfaces;
 using SFA.DAS.SharedOuterApi.Models;
 
 namespace SFA.DAS.EmployerIncentives.Application.Services
