+++ src/SFA.DAS.LevyTransferMatching.Api/AppStart/AddServiceRegistrationExtensions.cs
@@ -2,7 +2,7 @@
 using SFA.DAS.Api.Common.Infrastructure;
 using SFA.DAS.Api.Common.Interfaces;
 using SFA.DAS.LevyTransferMatching.Interfaces;
-using SFA.DAS.LevyTransferMatching.Services;
+using SFA.DAS.LevyTransferMatching.Application.Services;
 using SFA.DAS.SharedOuterApi.Configuration;
 using SFA.DAS.SharedOuterApi.Infrastructure;
 using SFA.DAS.SharedOuterApi.Interfaces;
