+++ src/SFA.DAS.SharedOuterApi/Interfaces/IAccountsApiClient.cs
@@ -1,6 +1,4 @@
-using SFA.DAS.SharedOuterApi.Interfaces;
-
-namespace SFA.DAS.EmployerIncentives.Interfaces
+namespace SFA.DAS.SharedOuterApi.Interfaces
 {
     public interface IAccountsApiClient<T> : IInternalApiClient<T>
     {
